# routes/api.php
<?php

declare(strict_types=1);

use App\Http\Controllers\AnalyticsController;
use App\Http\Controllers\Api\Admin\BrandController;
use App\Http\Controllers\Api\Admin\CategoryController;
use App\Http\Controllers\Api\AIController;
use App\Http\Controllers\Api\AuthController;
use App\Http\Controllers\Api\DocumentationController;
use App\Http\Controllers\Api\OrderController;
use App\Http\Controllers\Api\PriceSearchController;
use App\Http\Controllers\Api\ProductController;
use App\Http\Controllers\HealthController;
use App\Http\Controllers\PaymentController;
use App\Http\Controllers\PointsController;
use App\Http\Controllers\ReportController;
use App\Http\Controllers\ReviewController;
use App\Http\Controllers\SettingController;
use App\Http\Controllers\SystemController;
use App\Http\Controllers\UploadController;
use App\Http\Controllers\WebhookController;
use App\Models\PriceOffer;
use App\Models\Product;
use App\Models\Review;
use App\Models\User;
use App\Services\AIService;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Cache;
use Illuminate\Support\Facades\Http;
use Illuminate\Support\Facades\Route;
use Illuminate\Validation\ValidationException;

// Authentication routes with Rate Limiting
Route::post('/login', [AuthController::class, 'login'])->middleware('throttle:5,1');
Route::post('/register', [AuthController::class, 'register'])->middleware('throttle:3,1');
Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth:sanctum');
Route::middleware(['auth:sanctum', 'throttle:auth'])->get('/user', [AuthController::class, 'me']);
Route::middleware(['auth:sanctum', 'throttle:authenticated'])->get('/me', [AuthController::class, 'me']);

// Public API routes (no authentication required)
Route::middleware(['throttle:public'])->group(static function (): void {
    // Price search routes
    Route::get('/price-search', [PriceSearchController::class, 'search']);
    Route::get('/price-search/search', [PriceSearchController::class, 'search']);
    Route::get('/price-search/best-offer', [PriceSearchController::class, 'bestOffer']);
    Route::get('/price-search/supported-stores', [PriceSearchController::class, 'supportedStores']);

    // Public product routes
    Route::get('/products', [ProductController::class, 'index']);
    Route::get('/products/{id}', [ProductController::class, 'show'])->whereNumber('id');

    // Additional API routes for testing
    Route::get('/categories', static function () {
        return response()->json(['data' => [], 'message' => 'Categories endpoint']);
    });

    Route::get('/brands', static function () {
        return response()->json(['data' => [], 'message' => 'Brands endpoint']);
    });

    Route::get('/wishlist', static function () {
        return response()->json(['data' => [], 'message' => 'Wishlist endpoint']);
    });

    Route::get('/price-alerts', static function () {
        return response()->json(['data' => [], 'message' => 'Price alerts endpoint']);
    });

    Route::get('/reviews', static function () {
        return response()->json(['data' => [], 'message' => 'Reviews endpoint']);
    });

    Route::get('/search', static function () {
        return response()->json(['data' => [], 'message' => 'Search endpoint']);
    });

    Route::get('/ai', static function () {
        return response()->json(['data' => [], 'message' => 'AI endpoint']);
    });

    // Product creation requires authentication
    Route::post('/products', [ProductController::class, 'store']);
});

// Authenticated API routes
Route::middleware(['auth:sanctum', 'throttle:authenticated'])->group(static function (): void {
    // Protected product routes
    Route::put('/products/{id}', [ProductController::class, 'update']);
    Route::delete('/products/{id}', [ProductController::class, 'destroy']);

    // Reviews routes
    Route::post('/products/{product}/reviews', [ReviewController::class, 'store']);

    // Order routes
    Route::apiResource('orders', OrderController::class);

    // Secure upload route using UploadController
    Route::post('/uploads', [UploadController::class, 'store']);
});

// Product deletion requires authentication
// Route::middleware(['throttle:public'])->group(function () {
//     Route::delete('/products/{id}', [ProductController::class, 'destroy']);
// });

// Admin API routes (high rate limits)
// Use 'auth' guard to align with tests using actingAs without Sanctum
Route::middleware(['auth', 'admin', 'throttle:admin'])->group(static function (): void {
    // Admin-specific routes
    Route::get('/admin/stats', static function () {
        return response()->json([
            'uptime' => time() - strtotime('2025-01-01 00:00:00'),
            'total_users' => User::count(),
            'total_products' => Product::count(),
            'total_offers' => PriceOffer::count(),
            'total_reviews' => Review::count(),
            'active_users_today' => User::whereDate('created_at', today())->count(),
            'new_products_today' => Product::whereDate('created_at', today())->count(),
            'server_time' => now()->toISOString(),
            'status' => 'operational',
        ]);
    });

    // Admin resource routes
    Route::apiResource('admin/categories', CategoryController::class)->names('api.admin.categories');
    Route::apiResource('admin/brands', BrandController::class)->names('api.admin.brands');
});

// API Documentation (no rate limiting for documentation)
Route::get('/', [DocumentationController::class, 'index']);
Route::get('/documentation', [DocumentationController::class, 'index']);

// API Health check (comprehensive monitoring)
Route::get('/health', [HealthController::class, 'check'])->name('api.health.check');
Route::get('/health/ping', [HealthController::class, 'ping'])->name('api.health.ping');

// CSRF token route for testing - REMOVED FOR PRODUCTION
// Route::get('/csrf-token', function () {
//     return response()->json(['token' => uniqid('csrf_', true)]);
// });

// Debug route for best offer - REMOVED FOR PRODUCTION
// Route::get('/debug-best-offer', function (Request $request) {
//     return response()->json([
//         'message' => 'Debug route working',
//         'params' => $request->all(),
//         'url' => $request->url()
//     ]);
// });

// Simple test route
Route::get('/test-simple', static function () {
    return response()->json(['message' => 'Simple test route works']);
});

// Test route for API tests
Route::get('/test', static function () {
    return response()->json([
        'data' => ['message' => 'API test route works'],
        'status' => 'success',
    ]);
});

// POST route for validation testing
Route::post('/test', static function (Request $request) {
    try {
        $validated = $request->validate([
            'name' => 'required|string|max:255',
            'email' => 'required|email',
        ]);

        return response()->json(['message' => 'Validation passed', 'data' => $validated]);
    } catch (ValidationException $e) {
        return response()->json([
            'message' => 'Validation failed',
            'errors' => $e->errors(),
        ], 422);
    }
});

// Temporary best offer route outside middleware - test method call
Route::get('/best-offer-debug', static function (Request $request) {
    try {
        $controller = app(PriceSearchController::class);

        return $controller->bestOffer($request);
    } catch (Exception $e) {
        return response()->json([
            'error' => 'Method call failed',
            'message' => $e->getMessage(),
            'trace' => $e->getTraceAsString(),
        ], 500);
    }
});

// Direct test of the bestOffer method
Route::get('/direct-best-offer', static function (Request $request) {
    return response()->json([
        'message' => 'Direct route test',
        'params' => $request->all(),
        'method' => 'bestOffer',
        'controller' => 'PriceSearchController',
    ]);
});

// Versioned API routes
Route::prefix('v1')->middleware(['throttle:api'])->group(static function (): void {
    Route::get('/best-offer', [PriceSearchController::class, 'bestOffer']);
    Route::get('/supported-stores', [PriceSearchController::class, 'supportedStores']);
});

// Test API routes for external service testing
Route::middleware(['throttle:public'])->group(static function (): void {
    // AI Text Analysis API
    Route::post('/ai/analyze', [AIController::class, 'analyze']);

    // AI Product Classification API
    Route::post('/ai/classify-product', [AIController::class, 'classifyProduct']);

    Route::get('/external-data', static function () {
        try {
            $response = Http::get('https://api.external-service.com/data');

            return response()->json($response->json(), $response->status());
        } catch (Exception $e) {
            return response()->json(['error' => 'External service unavailable'], 503);
        }
    });

    Route::get('/slow-external-data', static function () {
        try {
            $response = Http::timeout(3)->get('https://api.slow-service.com/data');

            return response()->json($response->json(), $response->status());
        } catch (Exception $e) {
            return response()->json(['error' => 'Service timeout'], 408);
        }
    });

    Route::get('/error-external-data', static function () {
        try {
            $response = Http::get('https://api.error-service.com/data');
            if ($response->status() >= 400) {
                return response()->json(['error' => 'External service error'], 502);
            }

            return response()->json($response->json(), $response->status());
        } catch (Exception $e) {
            return response()->json(['error' => 'External service unavailable'], 503);
        }
    });

    Route::get('/authenticated-external-data', static function () {
        try {
            $response = Http::withHeaders([
                'Authorization' => 'Bearer test-token',
            ])->get('https://api.authenticated-service.com/data');

            return response()->json($response->json(), $response->status());
        } catch (Exception $e) {
            return response()->json(['error' => 'Authentication failed'], 401);
        }
    });

    Route::get('/rate-limited-external-data', static function () {
        try {
            $response = Http::get('https://api.rate-limited-service.com/data');
            if (429 === $response->status()) {
                return response()->json(['error' => 'Rate limited'], 429);
            }

            return response()->json($response->json(), $response->status());
        } catch (Exception $e) {
            return response()->json(['error' => 'Service unavailable'], 503);
        }
    });

    Route::get('/cached-external-data', static function () {
        return Cache::remember('external-data', 60, static function () {
            try {
                $response = Http::get('https://api.cacheable-service.com/data');

                return $response->json();
            } catch (Exception $e) {
                return ['error' => 'Service unavailable'];
            }
        });
    });

    Route::get('/fallback-external-data', static function () {
        try {
            // Try primary service first
            $response = Http::get('https://api.primary-service.com/data');
            if ($response->successful()) {
                return response()->json($response->json(), 200);
            }
        } catch (Exception $e) {
            // Primary service failed, try fallback
        }

        try {
            // Try fallback service
            $response = Http::get('https://api.fallback-service.com/data');

            return response()->json($response->json(), $response->status());
        } catch (Exception $e) {
            return response()->json(['error' => 'All services unavailable'], 503);
        }
    });
});

// Payment Routes
Route::middleware('auth:sanctum')->group(static function (): void {
    Route::get('/payment-methods', [PaymentController::class, 'getPaymentMethods']);
    Route::post('/orders/{order}/payments', [PaymentController::class, 'processPayment']);
    Route::post('/orders/{order}/refund', [PaymentController::class, 'refundPayment']);
});

// Order Routes
Route::middleware(['auth:sanctum', 'throttle:authenticated'])->group(static function (): void {
    Route::get('/orders', [OrderController::class, 'index']);
    Route::post('/orders', [OrderController::class, 'store']);
    Route::get('/orders/{order}', [OrderController::class, 'show']);
});

// Points & Rewards Routes
Route::middleware(['auth:sanctum', 'throttle:authenticated'])->group(static function (): void {
    Route::get('/points', [PointsController::class, 'index']);
    Route::post('/points/redeem', [PointsController::class, 'redeem']);
    Route::get('/rewards', [PointsController::class, 'getRewards']);
    Route::post('/rewards/{reward}/redeem', [PointsController::class, 'redeemReward']);
});

// Settings API routes
Route::middleware(['throttle:api'])->prefix('settings')->group(static function (): void {
    // Public read-only routes
    Route::get('/password-policy', [SettingController::class, 'getPasswordPolicySettings']);
    Route::get('/notifications', [SettingController::class, 'getNotificationSettings']);
    Route::get('/general', [SettingController::class, 'getGeneralSettings']);

    // Admin-only routes
    Route::middleware(['auth', 'admin'])->group(static function (): void {
        Route::get('/', [SettingController::class, 'index']);
        Route::put('/', [SettingController::class, 'update']);
        Route::get('/storage', [SettingController::class, 'getStorageSettings']);
        Route::get('/security', [SettingController::class, 'getSecuritySettings']);
        Route::get('/performance', [SettingController::class, 'getPerformanceSettings']);
        Route::post('/reset', [SettingController::class, 'resetToDefault']);
        Route::post('/import', [SettingController::class, 'importSettings']);
        Route::get('/export', [SettingController::class, 'exportSettings']);
        Route::get('/system-health', [SettingController::class, 'getSystemHealth']);
    });
});

// System API routes
Route::middleware(['throttle:api'])->prefix('system')->group(static function (): void {
    // Wrap system info in try/catch to return unified JSON on errors (as tests expect)
    Route::get('/info', static function () {
        try {
            return app(SystemController::class)->getSystemInfo();
        } catch (Throwable $e) {
            return response()->json([
                'success' => false,
                'message' => 'Failed to get system information',
            ], 500);
        }
    });

    // CRITICAL OPERATIONS - Require Admin Authentication
    Route::middleware(['auth', 'admin'])->group(static function (): void {
        Route::post('/migrations', [SystemController::class, 'runMigrations']);
        Route::post('/cache/clear', [SystemController::class, 'clearCache']);
        Route::post('/optimize', [SystemController::class, 'optimizeApp']);
        Route::post('/composer-update', [SystemController::class, 'runComposerUpdate']);
    });

    // Wrap performance metrics endpoint to return unified JSON on exceptions
    Route::get('/performance', static function () {
        try {
            return app(SystemController::class)->getPerformanceMetrics();
        } catch (Throwable $e) {
            return response()->json([
                'success' => false,
                'message' => 'Failed to get performance metrics',
            ], 500);
        }
    });
});

// Report API routes
Route::middleware(['throttle:api'])->prefix('reports')->group(static function (): void {
    // POST routes for generating reports
    Route::post('/product-performance', [ReportController::class, 'generateProductPerformanceReport']);
    Route::post('/user-activity', [ReportController::class, 'generateUserActivityReport']);
    Route::post('/sales', [ReportController::class, 'generateSalesReport']);
    Route::post('/custom', [ReportController::class, 'generateCustomReport']);
    Route::post('/export', [ReportController::class, 'exportReport']);

    // GET routes for retrieving reports
    Route::get('/system-overview', [ReportController::class, 'getSystemOverview']);
    Route::get('/engagement-metrics', [ReportController::class, 'getEngagementMetrics']);
    Route::get('/performance-metrics', [ReportController::class, 'getPerformanceMetrics']);
    Route::get('/top-stores', [ReportController::class, 'getTopStores']);
    Route::get('/price-trends', [ReportController::class, 'getPriceTrends']);
    Route::get('/most-viewed-products', [ReportController::class, 'getMostViewedProducts']);
});

// Analytics API routes
Route::middleware(['throttle:public'])->group(static function (): void {
    Route::get('/analytics/site', [AnalyticsController::class, 'siteAnalytics']);
});
Route::middleware(['auth:sanctum', 'throttle:authenticated'])->group(static function (): void {
    Route::get('/analytics/user', [AnalyticsController::class, 'userAnalytics']);
    Route::post('/analytics/track', [AnalyticsController::class, 'trackBehavior']);
});

// AI API routes
Route::middleware(['throttle:ai'])->prefix('ai')->group(static function (): void {
    Route::post('/analyze', static function (Request $request) {
        try {
            /** @var array{text: string, type: string} $validated */
            $validated = $request->validate([
                'text' => 'required|string|max:10000',
                'type' => 'required|string|in:general,product_analysis,product_classification,recommendations,sentiment',
            ]);

            $validTypes = ['general', 'product_analysis', 'product_classification', 'recommendations', 'sentiment'];
            if (! in_array($validated['type'], $validTypes, true)) {
                return response()->json([
                    'success' => false,
                    'error' => 'Validation failed',
                    'message' => 'Invalid input data',
                    'errors' => ['type' => ['The selected type is invalid.']],
                ], 422);
            }

            $aiService = app(AIService::class);
            $result = $aiService->analyzeText($validated['text'], $validated['type']);

            return response()->json([
                'success' => true,
                'data' => $result,
                'message' => 'Analysis completed successfully',
            ], 200);
        } catch (ValidationException $e) {
            return response()->json([
                'success' => false,
                'error' => 'Validation failed',
                'message' => 'Invalid input data',
                'errors' => $e->errors(),
            ], 422);
        } catch (Exception $e) {
            return response()->json([
                'success' => false,
                'error' => 'Analysis failed',
                'message' => 'فشل في تحليل النص',
                'details' => $e->getMessage(),
            ], 500);
        }
    });

    Route::post('/classify-product', static function (Request $request) {
        try {
            /** @var array{name: string, description: ?string, price: ?(float|int|string)} $validated */
            $validated = $request->validate([
                'name' => 'required|string|max:255',
                'description' => 'nullable|string|max:1000',
                'price' => 'nullable|numeric|min:0',
            ]);

            $aiService = app(AIService::class);
            $productDescription = $validated['description'] ?? '';
            $category = $aiService->classifyProduct($productDescription);

            return response()->json([
                'success' => true,
                'category' => $category,
                'confidence' => 0.8,
                'data' => [
                    'category' => $category,
                    'confidence' => 0.8,
                ],
                'message' => 'Product classified successfully',
            ], 200);
        } catch (ValidationException $e) {
            return response()->json([
                'success' => false,
                'error' => 'Validation failed',
                'message' => 'Invalid input data',
                'errors' => $e->errors(),
            ], 422);
        } catch (Exception $e) {
            return response()->json([
                'success' => false,
                'error' => 'Classification failed',
                'message' => 'فشل في تصنيف المنتج',
                'details' => $e->getMessage(),
            ], 500);
        }
    });

    Route::post('/analyze-image', static function (Request $request) {
        /** @var array{image_url: string} $validated */
        $validated = $request->validate([
            'image_url' => 'required|url|max:2048',
        ]);

        try {
            $aiService = app(AIService::class);
            $result = $aiService->analyzeImage($validated['image_url']);

            return response()->json([
                'success' => true,
                'data' => $result,
            ]);
        } catch (Exception $e) {
            return response()->json([
                'success' => false,
                'error' => 'فشل في تحليل الصورة',
                'message' => $e->getMessage(),
            ], 500);
        }
    });

    Route::post('/recommendations', static function (Request $request) {
        $validated = $request->validate([
            'preferences' => 'required|array|min:1',
            'products' => 'required|array|min:1',
            'products.*.name' => 'required|string|max:255',
            'products.*.description' => 'nullable|string|max:1000',
            'products.*.price' => 'nullable|numeric|min:0',
        ]);

        /*
         * @var array{
         *   preferences: array<string, mixed>,
         *   products: array<int, array<string, mixed>>
         * } $validated
         */
        try {
            $aiService = app(AIService::class);

            $recommendations = $aiService->generateRecommendations($validated['preferences'], $validated['products']);

            return response()->json([
                'success' => true,
                'recommendations' => $recommendations,
            ]);
        } catch (Exception $e) {
            return response()->json([
                'success' => false,
                'error' => 'فشل في توليد التوصيات',
                'message' => $e->getMessage(),
            ], 500);
        }
    });
});

// ============================================================================
// Webhook Routes (No Authentication - Verified by Signature)
// ============================================================================

Route::prefix('webhooks')->group(static function (): void {
    Route::post('/amazon', [WebhookController::class, 'amazon'])->name('webhooks.amazon');
    Route::post('/ebay', [WebhookController::class, 'ebay'])->name('webhooks.ebay');
    Route::post('/noon', [WebhookController::class, 'noon'])->name('webhooks.noon');
});

// Secure upload endpoint (authenticated), stores to private disk and returns signed URL
Route::post('/uploads', [UploadController::class, 'store'])
    ->middleware(['auth:sanctum', 'throttle:api'])
    ->name('uploads.store')
;