# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "COPRRA API",
        "description": "API for COPRRA - Price Comparison Platform",
        "contact": {
            "name": "COPRRA API Support",
            "email": "api@coprra.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.coprra.com",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        },
        {
            "url": "https://api.coprra.com/v2",
            "description": "Production Server v2"
        },
        {
            "url": "http://localhost:8000/api/v2",
            "description": "Development Server v2"
        }
    ],
    "paths": {
        "/products/{id}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update product",
                "description": "Update an existing product (Admin only)",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/products": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create product",
                "description": "Create a new product (Admin only)",
                "operationId": "createProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Brand": {
                "title": "Brand",
                "description": "Brand model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Apple"
                    },
                    "slug": {
                        "type": "string",
                        "example": "apple"
                    },
                    "logo": {
                        "type": "string",
                        "example": "https://example.com/logo.png",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ProductCreateRequest": {
                "title": "Product Create Request",
                "description": "Request data for creating a product",
                "required": [
                    "name",
                    "slug",
                    "price",
                    "category_id",
                    "brand_id"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "iPhone 15 Pro"
                    },
                    "slug": {
                        "type": "string",
                        "example": "iphone-15-pro"
                    },
                    "description": {
                        "type": "string",
                        "example": "Latest iPhone with advanced features"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 999.99
                    },
                    "image": {
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "brand_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "store_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "Product": {
                "title": "Product",
                "description": "Product model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "iPhone 15 Pro"
                    },
                    "slug": {
                        "type": "string",
                        "example": "iphone-15-pro"
                    },
                    "description": {
                        "type": "string",
                        "example": "Latest iPhone with advanced features"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 999.99
                    },
                    "image": {
                        "type": "string",
                        "example": "https://example.com/image.jpg",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "brand_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "store_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "brand": {
                        "$ref": "#/components/schemas/Brand"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "price_offers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PriceOffer"
                        }
                    }
                },
                "type": "object"
            },
            "ProductUpdateRequest": {
                "title": "Product Update Request",
                "description": "Request data for updating a product",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "iPhone 15 Pro Max"
                    },
                    "slug": {
                        "type": "string",
                        "example": "iphone-15-pro-max"
                    },
                    "description": {
                        "type": "string",
                        "example": "Updated description"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 1099.99
                    },
                    "image": {
                        "type": "string",
                        "example": "https://example.com/new-image.jpg"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": false
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "brand_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "store_id": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "apiKey": {
                "type": "apiKey",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
}